v1.1.1 release
RTP detection change, it now only checks that target block is open and that beneath target block is not water. Should have a higher success rate.

Got rid of some souting used for debugs

fixed start not clearing achieves if a non-op issued it.

Settings now save properly, and fixed server crashing when attempting to start a game with only 1 player online.

the tab completer got pruned a bit to be prettier.

anyStart got added to the saver.

v1.1.1 of the plugin is now included in the git